+++ CHANGES.rst
@@ -7,7 +7,8 @@ Changelog
 0.4 (unreleased)
 ~~~~~~~~~~~~~~~~
 
-* Add ``Request.base_url``.  (Issue #4)
+* Add ``Request.host_url``, ``Request.script_name`` and
+  ``Request.get_url(path, full=False)``.  (Issue #4)
 
 
 0.3 (2013-06-12)